Send seek as long

+++ AudioPlayer/Source/State/StatePaused.swift
@@ -26,8 +26,8 @@ class StatePaused:StateProtocol {
         player.state = States.stopped
     }
     
-    func seek(player:Player, seconds:Float) throws {
-        player.provider.seek(seconds:Int(seconds))
+    func seek(player:Player, seconds:Int) throws {
+        player.provider.seek(seconds:seconds)
     }
     
     func pause(player:Player) throws { throw PlayerError.canNotPause }
