+++ AudioPlayer/Source/State/StateProtocol+Default.swift
@@ -5,5 +5,5 @@ extension StateProtocol {
     func play(player:Player) throws { }
     func pause(player:Player) throws { }
     func stop(player:Player) throws { }
-    func seek(player:Player, seconds:Float) throws { }
+    func seek(player:Player, seconds:Int) throws { }
 }
